mrquest

+++ .config/fish/functions/mrquest.fish
@@ -1,5 +1,8 @@
 # Defined interactively
 function mrquest
-    mv ~/quest/Movies/* ~/d/68_Vr_Tv_Keep/
-    fd --max-results=3 . ~/d/68_Vr_Tv/_new/ | xargs -I{} mv {} ~/quest/Movies/
+    ls ~/quest/Movies/
+    if confirm
+        mv ~/quest/Movies/* ~/d/68_Vr_Tv_Keep/
+        fd --max-results=3 . ~/d/68_Vr_Tv/_new/ | xargs -I{} mv {} ~/quest/Movies/
+    end
 end
